+++ modules/fetchAndRenderComments.js
@@ -1,19 +1,27 @@
 import { getAllComments } from "./api.js";
-import { getTime } from "./utilitities.js";
+import { getTime, sanitizeHtml } from "./utilitities.js";
 import { renderComments } from "./renderElements.js";
+import { renderInputBox } from "./renderElements.js";
 
 export function fetchAndRenderComments() {
 
+    const addFormElement = document.querySelector('.add-form');
     const commentLoaderElement = document.querySelector('.comment-loader');
+
+    addFormElement.innerHTML = `
+            <div class="comment-add-container">
+                <p>Комментарий добавляется...</p>
+                <img src="./spinner.svg" class="spinner">
+            </div>
+        `;
 
     getAllComments()
         .then((responseData) => {
             let comments = responseData.comments.map((comment) => {
                 return {
-                    name: comment.author.name,
+                    name: sanitizeHtml(comment.author.name),
                     date: getTime(new Date(comment.date)),
-                    text: comment.text.replaceAll("<", "&lt;").replaceAll(">", "&gt;")
-                        .replaceAll('QUOTE_BEGIN', '<div class="quote">').replaceAll('QUOTE_END', '</div>'),
+                    text: sanitizeHtml(comment.text),
                     isLiked: comment.isLiked,
                     likes: comment.likes
                 }
@@ -21,6 +29,9 @@ export function fetchAndRenderComments() {
             commentLoaderElement.classList.add('hide-comment-loader');
             renderComments({ comments });
             return true;
+        })
+        .then(() => {
+            renderInputBox("", "");
         })
         .catch((error) => {
             console.warn(error);
